Create Functions file
