+++ backend/Features/Profiles/ProfilesController.cs
@@ -15,7 +15,7 @@ namespace Backend.Features.Profiles;
 [ApiController]
 [Route("api/profiles")]
 [Authorize]
-public sealed class ProfilesController(AppDbContext db, IBlobStorageService blobStorage, ILogger<ProfilesController> logger) : ControllerBase
+public sealed partial class ProfilesController(AppDbContext db, IBlobStorageService blobStorage, ILogger<ProfilesController> logger) : ControllerBase
 {
     /// <summary>
     /// Get a public profile by ID, respecting privacy settings.
@@ -396,15 +396,6 @@ public sealed class ProfilesController(AppDbContext db, IBlobStorageService blob
         return Ok(response);
     }
 
-    private static readonly Dictionary<string, string> AllowedMimeTypes = new(StringComparer.OrdinalIgnoreCase)
-    {
-        ["image/jpeg"] = "jpg",
-        ["image/png"] = "png",
-        ["image/webp"] = "webp",
-    };
-
-    private const long MaxPhotoSizeBytes = 5 * 1024 * 1024;
-
     /// <summary>
     /// Upload or replace the current authenticated user's profile photo.
     /// </summary>
@@ -441,7 +432,7 @@ public sealed class ProfilesController(AppDbContext db, IBlobStorageService blob
             return ValidationProblem(ModelState);
         }
 
-        if (!AllowedMimeTypes.TryGetValue(photo.ContentType, out var fileExtension))
+        if (!_allowedMimeTypes.TryGetValue(photo.ContentType, out var fileExtension))
         {
             ModelState.AddModelError(nameof(photo), "Only JPEG, PNG, and WebP images are accepted.");
             return ValidationProblem(ModelState);
@@ -538,44 +529,5 @@ public sealed class ProfilesController(AppDbContext db, IBlobStorageService blob
         await blobStorage.DeleteBlobByUrlAsync(oldPhotoUrl, cancellationToken);
 
         return NoContent();
-    }
-
-    private bool TryBuildLocation(double? latitude, double? longitude, out Point? location)
-    {
-        location = null;
-
-        if (latitude.HasValue != longitude.HasValue)
-        {
-            ModelState.AddModelError(nameof(UpdateOwnProfileRequest.Latitude), "Both Latitude and Longitude must be provided together.");
-            ModelState.AddModelError(nameof(UpdateOwnProfileRequest.Longitude), "Both Latitude and Longitude must be provided together.");
-            return false;
-        }
-
-        if (!latitude.HasValue || !longitude.HasValue)
-        {
-            return true;
-        }
-
-        if (!double.IsFinite(latitude.Value) || latitude.Value is < -90 or > 90)
-        {
-            ModelState.AddModelError(nameof(UpdateOwnProfileRequest.Latitude), "Latitude must be between -90 and 90.");
-            return false;
-        }
-
-        if (!double.IsFinite(longitude.Value) || longitude.Value is < -180 or > 180)
-        {
-            ModelState.AddModelError(nameof(UpdateOwnProfileRequest.Longitude), "Longitude must be between -180 and 180.");
-            return false;
-        }
-
-        location = new Point(longitude.Value, latitude.Value) { SRID = 4326 };
-        return true;
-    }
-
-    private Task<bool> ProfileExistsAsync(Guid id, CancellationToken cancellationToken)
-    {
-        return db.Profiles
-            .AsNoTracking()
-            .AnyAsync(profile => profile.Id == id, cancellationToken);
     }
 }
